[expo][3.0-release-test] some fixes
**Summary**

- fix firebase analytics version consistency.
- moved push notif token sync loigc to UserRoot instead of AppRoot as we don't want show the dialog during they accept policies.

**Test**

- expo

**Issue**

- N/A
