+++ expo/package.json
@@ -40,7 +40,7 @@
     "@babel/plugin-proposal-class-properties": "^7.18.6",
     "@babel/plugin-proposal-private-property-in-object": "^7.21.0",
     "@react-native-async-storage/async-storage": "1.23.1",
-    "@react-native-firebase/analytics": "^21.6.1",
+    "@react-native-firebase/analytics": "17.4.3",
     "@react-native-firebase/app": "^17.4.3",
     "@react-native-firebase/app-check": "^17.4.3",
     "@react-native-firebase/auth": "^17.4.3",
